Fix lint findings from the status-page change
Three issues the linter caught after the first commit:

- liveAudioStatus: the hook reset state to "loading" from inside the effect body,
  which cascades renders. The result now carries the docId it describes and the
  hook derives "loading" for a mismatch, so switching sessions needs no reset.
- StatusView: ages were computed with Date.now() during render (impure, and the
  displayed number drifted between polls). The hook now stamps receivedAt when a
  payload lands and ages are measured against that.
- Moved the tile derivation into src/statusTiles.ts so StatusView.tsx exports only
  components (react-refresh), which also makes the tone rules directly testable.

Also fixes a test that asserted on 'ES' where the upper-casing is CSS only.

npm run lint, npm run typecheck, npm test (318 passed), and uv run pytest all pass.
